bugfix - minCount should apply to both words and labels
In the paper (https://arxiv.org/pdf/1607.01759v3.pdf) in section 3.2, the authors describe limiting the words and labels they use to those with a frequency > 100. The code above previously only performs this limit on words with this threshold, allowing labels even if they are only seen once. The fix above alleviates this problem allowing individuals to repeat the YFCC100M experiment. 

Without the bugfix, the YFCC100M experiment tries to keep 7213668 labels
With the bugfix, the YFCC100M experiment instead trains on 312116 labels.